Refactoring : Json Key Front 와 동일 작업 수행

+++ src/main/java/com/sosim/server/group/GroupService.java
@@ -72,6 +72,11 @@ public class GroupService {
             throw new CustomException(CodeType.NONE_ADMIN);
         }
 
+        if (participantService.getCountParticipantAtGroup(groupEntity.getId()) > 1) {
+            throw new CustomException(CodeType.NONE_ZERO_PARTICIPANT);
+        }
+
+        participantService.deleteParticipantEntity(userService.getUser(userId), getGroupEntity(groupId));
         groupEntity.delete();
     }
